requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-nvidia-ai-endpoints
pydantic
python-dotenv
pytest